Added gethrtime and changed modelica_integer to long
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@748 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ mosh/src/value.cpp
@@ -47,7 +47,7 @@ value::value(bool val)
   m_basic_type = create_boolean();
 }
 
-value::value(int val)
+value::value(long val)
 {
   m_integer = val;
   m_basic_type = create_integer();
@@ -182,7 +182,7 @@ void value::set_value(std::string val)
    m_basic_type = create_string();
 }
 
-void value::set_value(int val)
+void value::set_value(long val)
 {
   m_integer = val;
    m_basic_type = create_integer();
@@ -1423,7 +1423,7 @@ value read_result_file(const char* filename)
 		  break; 
 		}
 	      read_to_eol(fp);
-	      vals.push_back(value(i));
+	      vals.push_back(value((long)i));
 	    }
 	  else 
 	    {
